@@ -4,7 +4,7 @@
     <Version>2019.8.24.1234</Version>
     <AssemblyVersion>2019.8.24.1234</AssemblyVersion>
     <FileVersion>2019.8.24.1234</FileVersion>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <LangVersion>latest</LangVersion>
     <Configurations>Debug;Release;Tool</Configurations>
   </PropertyGroup>
@@ -13,7 +13,7 @@
     <WarningsAsErrors />
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.TestHost" Version="2.2.0" />
+    <PackageReference Include="Microsoft.AspNetCore.TestHost" Version="3.0.0" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\HttPlaceholder\HttPlaceholder.csproj" />
