@@ -1,66 +1,68 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <PackageId>HttPlaceholder</PackageId>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <Version>2019.8.24.1234</Version>
-    <AssemblyVersion>2019.8.24.1234</AssemblyVersion>
-    <FileVersion>2019.8.24.1234</FileVersion>
-    <RuntimeIdentifiers>win81-x64;win10-x64;ubuntu.14.04-x64;ubuntu.16.04-x64</RuntimeIdentifiers>
-    <LangVersion>latest</LangVersion>
-    <Authors>Duco Winterwerp</Authors>
-    <Company>Ducode</Company>
-    <RepositoryUrl>https://github.com/dukeofharen/httplaceholder</RepositoryUrl>
-    <Description>A very flexible cross platform HTTP stub application.</Description>
-    <IsPackable>true</IsPackable>
-    <PackAsTool>false</PackAsTool>
-    <ToolCommandName>httplaceholder</ToolCommandName>
-    <Configurations>Debug;Release;Tool</Configurations>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <WarningsAsErrors />
-    <NoWarn>1701;1702;RCS1138;CS1591</NoWarn>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <WarningsAsErrors />
-    <NoWarn>1701;1702; CS1591</NoWarn>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Tool|AnyCPU'">
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <WarningsAsErrors />
-    <NoWarn>1701;1702; CS1591</NoWarn>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" Version="2.2.7" />
-    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="2.2.0" />
-    <PackageReference Include="NSwag.AspNetCore" Version="13.0.6" />
-    <PackageReference Include="YamlDotNet" Version="6.1.2" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\HttPlaceholder.Application\HttPlaceholder.Application.csproj" />
-    <ProjectReference Include="..\HttPlaceholder.Common\HttPlaceholder.Common.csproj" />
-    <ProjectReference Include="..\HttPlaceholder.Infrastructure\HttPlaceholder.Infrastructure.csproj" />
-    <ProjectReference Include="..\HttPlaceholder.Persistence\HttPlaceholder.Persistence.csproj" />
-  </ItemGroup>
-  <ItemGroup Condition="'$(Configuration)|$(Platform)'!='Tool|AnyCPU'">
-    <Content Update="_config.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-  </ItemGroup>
-  <ItemGroup>
-    <Content Update="gui\*.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <None Update="gui\**">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-  <ItemGroup>
-    <None Update="key.pfx">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <PackageId>HttPlaceholder</PackageId>
+        <TargetFramework>netcoreapp3.0</TargetFramework>
+        <PublishTrimmed>true</PublishTrimmed>
+        <GenerateDocumentationFile>true</GenerateDocumentationFile>
+        <Version>2019.8.24.1234</Version>
+        <AssemblyVersion>2019.8.24.1234</AssemblyVersion>
+        <FileVersion>2019.8.24.1234</FileVersion>
+        <RuntimeIdentifiers>win81-x64;win10-x64;ubuntu.14.04-x64;ubuntu.16.04-x64</RuntimeIdentifiers>
+        <LangVersion>latest</LangVersion>
+        <Authors>Duco Winterwerp</Authors>
+        <Company>Ducode</Company>
+        <RepositoryUrl>https://github.com/dukeofharen/httplaceholder</RepositoryUrl>
+        <Description>A very flexible cross platform HTTP stub application.</Description>
+        <IsPackable>true</IsPackable>
+        <PackAsTool>false</PackAsTool>
+        <ToolCommandName>httplaceholder</ToolCommandName>
+        <Configurations>Debug;Release;Tool</Configurations>
+    </PropertyGroup>
+    <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+        <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+        <WarningsAsErrors />
+        <NoWarn>1701;1702;RCS1138;CS1591</NoWarn>
+    </PropertyGroup>
+    <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
+        <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+        <WarningsAsErrors />
+        <NoWarn>1701;1702; CS1591</NoWarn>
+    </PropertyGroup>
+    <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Tool|AnyCPU'">
+        <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+        <WarningsAsErrors />
+        <NoWarn>1701;1702; CS1591</NoWarn>
+    </PropertyGroup>
+    <ItemGroup>
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.0.0" />
+        <PackageReference Include="Microsoft.AspNetCore.SignalR.Protocols.NewtonsoftJson" Version="3.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="3.0.0" />
+        <PackageReference Include="NSwag.AspNetCore" Version="13.0.6" />
+        <PackageReference Include="YamlDotNet" Version="6.1.2" />
+    </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\HttPlaceholder.Application\HttPlaceholder.Application.csproj" />
+        <ProjectReference Include="..\HttPlaceholder.Common\HttPlaceholder.Common.csproj" />
+        <ProjectReference Include="..\HttPlaceholder.Infrastructure\HttPlaceholder.Infrastructure.csproj" />
+        <ProjectReference Include="..\HttPlaceholder.Persistence\HttPlaceholder.Persistence.csproj" />
+    </ItemGroup>
+    <ItemGroup Condition="'$(Configuration)|$(Platform)'!='Tool|AnyCPU'">
+        <Content Update="_config.json">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </Content>
+    </ItemGroup>
+    <ItemGroup>
+        <Content Update="gui\*.json">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </Content>
+        <None Update="gui\**">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+    </ItemGroup>
+    <ItemGroup>
+        <None Update="key.pfx">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+    </ItemGroup>
 </Project>
